v.2.8a
Banco de dados reestruturado: /bd agora é /server, pastas modificadas e simplificadas,
sem mais precisar executar o código duas vezes.
